requirements.txt
streamlit
openai
python-dotenv
chromadb
langchain
sentence-transformers